zigbee: changed the java types of some entities (eg. command ID) and updated some javadocs accordingly.

+++ org.osgi.service.zigbee/src/org/osgi/service/zigbee/ZCLHeader.java
@@ -28,7 +28,7 @@ public interface ZCLHeader {
 	 * 
 	 * @return the command identifier of this frame.
 	 */
-	int getCommandId();
+	short getCommandId();
 
 	/**
 	 * Returns the manufacturer code of this frame.
